[Not thought out] Fix for crashing of plugin?
Now, in OnDisable, we don't do anything. I don't know the rational behind Shutdown so it may have been justified, but it was causing crashes due to its interaction with an awoken haptics explorer. Now EnsurePluginIsValid just tries to access the singleton, which internally should instantiate the plugin. In HardlightManager I'm now doing InitPluginIfNull when the first instance is created only, hopefully